requirements.txt
a2wsgi==1.7.0
annotated-types==0.5.0
anyio==4.0.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==3.4.8
fastapi==0.103.0
fastapi-jwt-auth==0.5.0
h11==0.14.0
httptools==0.6.0
idna==3.4
passlib==1.7.4
psycopg2==2.9.7
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.12
pydantic_core==2.6.3
PyJWT==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
wfastcgi==3.0.0
